continuing to add in new enhancements and adding in code so enemies can use bonuses as well as the player

+++ src/net/anorrah/damageObject.java
@@ -2,11 +2,11 @@ package net.anorrah;
 
 public class damageObject {
 	
-	public double damage;
+	public int damage;
 	public enum Type {physical, fire, ice, lightning, dark};
 	public Type type;
 	
-	public damageObject(double damage, Type type)
+	public damageObject(int damage, Type type)
 	{
 		this.damage = damage;
 		this.type = type;
